requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
python-dotenv>=1.0.0

# PDF processing
pdfplumber>=0.10.0
pymupdf>=1.23.0

# Data validation
pydantic>=2.5.0
jsonschema>=4.20.0

# CLI
click>=8.1.7

# Utilities
requests>=2.31.0
pyyaml>=6.0.0
